Unit Tests of utils And refactoring
the whole utils and photos packages are tested.Refactored Credit Card saver module and utils classes.

+++ src/main/java/com/webshop/webapp/entity/service/CreditCardServiceImpl.java
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.webshop.webapp.entity.CreditCard;
+import com.webshop.webapp.entity.ForbiddenAction;
+import com.webshop.webapp.utils.IsThisCardRelatedToOngoingOrderChecker;
 import com.webshop.webapp.webservice.CreditCardWebservice;
 
 @Service
@@ -13,6 +15,9 @@ public class CreditCardServiceImpl implements CreditCardService{
 
 	@Autowired
 	private CreditCardWebservice creditCardWebservice;
+	
+	@Autowired
+	private IsThisCardRelatedToOngoingOrderChecker isThisCardRelatedToOngoingOrderChecker;
 	
 	@Override
 	public CreditCard getCreditCardById(int creditCardId) {
@@ -30,8 +35,8 @@ public class CreditCardServiceImpl implements CreditCardService{
 	}
 
 	@Override
-	public void save(CreditCard creditCard) {
-		creditCardWebservice.save(creditCard);
+	public CreditCard save(CreditCard creditCard) {
+		return creditCardWebservice.save(creditCard);
 		
 	}
 
@@ -43,7 +48,12 @@ public class CreditCardServiceImpl implements CreditCardService{
 
 	@Override
 	public void removeCreditCardById(int creditCardId) {
-		creditCardWebservice.removeCreditCardById(creditCardId);
+		
+		if(!isThisCardRelatedToOngoingOrderChecker.check(creditCardId)) 
+			creditCardWebservice.removeCreditCardById(creditCardId);
+		else
+			throw new ForbiddenAction();
+		
 	}
 
 }
